Arreglado todos los errores de la actividad añadir venta

+++ ProyectoFinal/app/src/main/java/clases/Albaran.kt
@@ -19,7 +19,7 @@ data class Albaran(
     @ColumnInfo("nombreCliente") var nombreCliente: String?,
     @ColumnInfo("fecha") var fecha: LocalDate?,
     @ColumnInfo("estado") var estado: String?,
-    @ColumnInfo("precio") var precioTotal: Int = 0,
+    @ColumnInfo("precio") var precioTotal: Float = 0f,
     @ColumnInfo("dni") var cliente: String
     ) : Parcelable
