+++ ProyectoFinal/app/src/main/java/clases/Factura_Producto.kt
@@ -12,7 +12,7 @@ data class Factura_Producto(
     @PrimaryKey(autoGenerate = true) @ColumnInfo("id") var referencia: Int=0,
     @ColumnInfo("tituloFactura") var tituloFactura: String,
     @ColumnInfo("nombreProducto") var nombreProducto: String?,
-    @ColumnInfo("precio") var precio: Int = 0,
+    @ColumnInfo("precio") var precio: Float = 0f,
     @ColumnInfo("cantidad") var cantidad: Int = 0,
-    @ColumnInfo("total") var total: Int = 0
+    @ColumnInfo("total") var total: Float = 0f
 ) : Parcelable
